app/requirements.txt
asgiref==3.5.0
attrs==21.4.0
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-bulk==0.2.1
drf-spectacular==0.21.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.4.0
psycopg2==2.9.3
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
